Separate operator implementations
